+++ src/dpr/components/_filters/utils.ts
@@ -1,11 +1,12 @@
 import { Request } from 'express'
-import { DateFilterValue, DateRange, FilterValue } from '../filters/types'
 import { FilterType } from './filter-input/enum'
+import type { components } from '../../types/api'
+import type { FilterOption } from './filter-input/types'
+import type { DateFilterValue, DateRange, FilterValue } from '../filters/types'
 
 import SelectedFiltersUtils from './filters-selected/utils'
-import AsyncFiltersUtils from '../_async/async-filters-form/utils'
-import { RenderFiltersReturnValue } from '../_async/async-filters-form/types'
-import { components } from '../../types/api'
+import DateRangeInputUtils from '../_inputs/date-range/utils'
+import DateInputUtils from '../_inputs/date-input/utils'
 
 const setFilterValuesFromRequest = (filters: FilterValue[], req: Request, prefix = 'filters.'): FilterValue[] => {
   const { preventDefault } = req.query
@@ -56,6 +57,91 @@ const handleDateValue = (filter: FilterValue, req: Request, prefix: string) => {
   return dateValue || min || max || '1977-05-25'
 }
 
+/**
+ * Initialises the filters from the definition
+ *
+ * @param {components['schemas']['VariantDefinition']} definition
+ * @return {*}
+ */
+const getFiltersFromDefinition = (fields: components['schemas']['FieldDefinition'][], interactive = false) => {
+  return fields
+    .filter((f) => f.filter)
+    .filter((f) => {
+      // TODO: fix types here once it has been defined.
+      return interactive
+        ? (<components['schemas']['FieldDefinition'] & { interactive?: boolean }>f.filter).interactive
+        : !(<components['schemas']['FieldDefinition'] & { interactive?: boolean }>f.filter).interactive
+    })
+    .map((f) => {
+      const { display: text, name, filter } = f
+      const { type, staticOptions, dynamicOptions, defaultValue, mandatory, pattern } = filter
+
+      const options: FilterOption[] = []
+      if (staticOptions) {
+        if (type === FilterType.select) {
+          options.push({ value: 'no-filter', text: 'Select your option', disabled: true, selected: true })
+        }
+        staticOptions.forEach((o) => {
+          options.push({ value: o.name, text: o.display })
+        })
+      }
+
+      let filterData: FilterValue = {
+        text,
+        name,
+        type: type as FilterType,
+        options: options.length ? options : null,
+        value: defaultValue || null,
+        minimumLength: dynamicOptions ? dynamicOptions.minimumLength : null,
+        dynamicResourceEndpoint: null,
+        mandatory,
+        pattern,
+      }
+
+      if (type === FilterType.dateRange.toLowerCase()) {
+        const dateRegEx = /^\d{1,4}-\d{1,2}-\d{2,2} - \d{1,4}-\d{1,2}-\d{1,2}$/
+        const { min, max } = filter
+        let startValue
+        let endValue
+
+        if (min) startValue = min
+        if (max) endValue = max
+
+        let value
+        if (defaultValue && defaultValue.match(dateRegEx)) {
+          ;[startValue, endValue] = defaultValue.split(' - ')
+          value = DateRangeInputUtils.setDateRangeValuesWithinMinMax(filter, startValue, endValue)
+        } else if (defaultValue) {
+          value = defaultValue
+        } else {
+          value = DateRangeInputUtils.setDateRangeValuesWithinMinMax(filter, startValue, endValue)
+        }
+
+        filterData = filterData as unknown as DateFilterValue
+        filterData = {
+          ...filterData,
+          min: filter.min,
+          max: filter.max,
+          value,
+        }
+
+        filterData.relativeOptions = DateRangeInputUtils.getRelativeDateOptions(filterData.min, filterData.max)
+      }
+
+      if (type === FilterType.date.toLowerCase()) {
+        filterData = filterData as unknown as DateFilterValue
+        filterData = {
+          ...filterData,
+          value: DateInputUtils.setDateValueWithinMinMax(filter),
+          min: filter.min,
+          max: filter.max,
+        }
+      }
+
+      return filterData
+    })
+}
+
 const getFilters = async ({
   fields,
   req,
@@ -67,8 +153,8 @@ const getFilters = async ({
   interactive?: boolean
   prefix?: string
 }) => {
-  const defaultFilterData = <RenderFiltersReturnValue>await AsyncFiltersUtils.renderFilters(fields, interactive)
-  const filters = setFilterValuesFromRequest(defaultFilterData.filters, req)
+  const defaultFilters = await getFiltersFromDefinition(fields, interactive)
+  const filters = setFilterValuesFromRequest(defaultFilters, req)
   const selectedFilters = SelectedFiltersUtils.getSelectedFilters(filters, prefix)
   return {
     filters,
@@ -77,6 +163,7 @@ const getFilters = async ({
 }
 
 export default {
+  getFiltersFromDefinition,
   setFilterValuesFromRequest,
   getFilters,
 }
